+++ .vscode/settings.json
@@ -9,6 +9,7 @@
         "ignoretests",
         "kisielk",
         "sbom",
-        "SPDXID"
+        "SPDXID",
+        "structs"
     ]
 }
