+++ cyclonedx/cyclonedx_test.go
@@ -1,0 +1,14 @@
+package cyclonedx
+
+import (
+	"testing"
+
+	cyclone "github.com/CycloneDX/cyclonedx-go"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestToSPDX(t *testing.T) {
+	c := cyclone.NewBOM()
+	bom := ToSPDX(c)
+	assert.NotNil(t, bom)
+}
