+++ go.mod
@@ -2,7 +2,10 @@ module github.com/devops-kung-fu/bomber
 
 go 1.18
 
-require github.com/stretchr/testify v1.8.0
+require (
+	github.com/CycloneDX/cyclonedx-go v0.6.0
+	github.com/stretchr/testify v1.8.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
